// src/modules/country-list-with-dial-code-and-flag/types.ts
interface CountryInterface {
  name: string
  dial_code: string
  code: string
  flag: string
  preferred?: boolean
  secondary?: boolean
  area_codes?: Array<string>
  country_code?: string
}

interface CountryFlagSvgInterface {
  [key: string]: string
}

interface FilterOption {
  withSecondary: boolean
}

export type { CountryInterface, CountryFlagSvgInterface, FilterOption }